separa misTorneos en varios componentes
