+++ frontend/src/pages/Torneos/Crear.jsx
@@ -1,0 +1,72 @@
+import axios from "axios";
+import { useState } from "react";
+import * as React from 'react';
+
+export default function Crear() {
+
+    const [datos, setDatos] = useState({
+        nombre: ''
+    })
+             const handleInputChange = (e) => {
+                setDatos({
+                    ...datos,
+                    [e.target.name] : e.target.value
+                })
+            }
+            const enviarDatos = (event) => {
+                event.preventDefault()
+                axios
+                .post(`http://127.0.0.1:3001/api/torneos/`, {
+                  Nombre: datos.nombre,
+                  FechaCreacion: localStorage.Matchday,
+                  UsuarioCreador: localStorage.userID
+                })
+                .then((res) => {
+                  const torneoID = res.data.id
+                  axios
+                    .post(`http://127.0.0.1:3001/api/rankings/`, {
+                        Puntos: 0,
+                        TorneoID: torneoID, //el recien creado
+                        UsuarioID: localStorage.userID
+                    })
+                    .then((res) => {
+
+                    })
+                    .catch(() => console.log("error")); 
+                    window.location.reload(true);
+                    window.alert(
+                        `Creado con exito`
+                        );
+                })
+                .catch(() => console.log("error"));
+            }
+
+        return (
+            <div className="w-100 d-flex flex-column align-items-left">
+              <div className="card d-flex flex-column">
+                <form className="row" onSubmit={enviarDatos}>
+                <div className="card-body">
+                <h4 className="card-title">Crear un torneo privado</h4>
+                <label for="nombre">Nombre del torneo:</label>
+                <input className="form-control" type="text" placeholder="Nombre" onChange={handleInputChange} name="nombre"></input>
+                <label for="liga">Liga:</label>
+                <br></br>
+                <select className="custom-select my-1 mr-2" id="inlineFormCustomSelect" disabled>
+                <option selected>Champions League</option>
+                </select>
+                <div className="form-group row">
+                <div className="col-sm-10">
+                <button type="submit" className="btn btn-primary mt-2" onSubmit="console.log(ok)">Aceptar</button>
+                </div>
+                </div>
+             </div>
+                </form>
+             </div>
+             
+            </div>
+          );
+    
+  };
+  
+
+
